requirements.txt
fastapi
requests
pytest
pandas
uvicorn
pyyaml
pydantic
openai
google-auth
google-apps-chat
